[REFACTOR] : LeftAlignedCollectionViewFlowLayout
