+++ MLS/MLS/SharedViews/LeftAlignedCollectionViewFlowLayout.swift
@@ -1,0 +1,30 @@
+//
+//  LeftAlignedCollectionViewFlowLayout.swift
+//  MLS
+//
+//  Created by JINHUN CHOI on 2024/03/12.
+//
+
+import UIKit
+
+class LeftAlignedCollectionViewFlowLayout: UICollectionViewFlowLayout {
+    override func layoutAttributesForElements(in rect: CGRect) -> [UICollectionViewLayoutAttributes]? {
+        let attributes = super.layoutAttributesForElements(in: rect)
+
+        var leftMargin = sectionInset.left
+        var maxY: CGFloat = -1.0
+        attributes?.forEach { layoutAttribute in
+            if layoutAttribute.frame.origin.y >= maxY {
+                leftMargin = sectionInset.left
+            }
+
+            layoutAttribute.frame.origin.x = leftMargin
+
+            leftMargin += layoutAttribute.frame.width + minimumInteritemSpacing
+            maxY = max(layoutAttribute.frame.maxY, maxY)
+        }
+
+        estimatedItemSize = UICollectionViewFlowLayout.automaticSize
+        return attributes
+    }
+}
